solution/solution-ecs-sts-token/code-example/python/sls_sdk/requirements.txt
alibabacloud-credentials>=0.3.5
aliyun-log-python-sdk>=0.9.3